+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Function `array_equals()` that checks if both arrays have the same keys and their values are equal
+
 ## [1.1.1] - 2022-05-27
 
 ### Added